pyproject.toml
[project]
name = "immiclaw-test"
version = "0.1.0"
description = "LLM-driven Playwright web testing tool"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27.0",
    "openai>=2.29.0",
    "playwright>=1.58.0",
    "pydantic>=2.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["immiclaw_test"]
force-include = { "main.py" = "main.py" }

[project.scripts]
immiclaw-test = "main:main"
llmtest = "immiclaw_test.llm_proxy.full_test:main"

[tool.hatch.build.targets.sdist]
include = ["immiclaw_test", "main.py", "README.md", "config"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100
exclude = [
    "config/scenarios.legacy-immiclaw-2026-03-21",
]

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
